feat: move and organise files to modules and utils
