+++ src/index.ts
@@ -1,1 +1,14 @@
-export * from './request';
+/*
+* Request module
+*/
+export * from './modules/request';
+
+/*
+* Data module
+*/
+export * from './modules/data';
+
+/*
+* Utils
+*/
+export * from './utils';
