+++ src/modules/request/index.ts
@@ -9,7 +9,6 @@ const PATCH = 'patch';
 const DELETE = 'delete';
 
 export const METHODS = { GET, POST, PUT, DELETE, PATCH };
-
 /*
  * This is a factory for requests. It will create a new instance that will provide all the necessary tools
  * for making requests and add or remove tokens to your requests
